Replace credit scoring example for EU AI Act clarity
