+++ kogito-quarkus-examples/dmn-pmml-quarkus-example/src/test/java/org/kie/kogito/dmn/pmml/quarkus/example/DMNEquipmentMaintenanceTest.java
@@ -16,50 +16,44 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.dmn.pmml.springboot.example;
+package org.kie.kogito.dmn.pmml.quarkus.example;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.annotation.DirtiesContext;
 
+import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.path.json.config.JsonPathConfig;
 
 import static io.restassured.RestAssured.given;
 import static io.restassured.config.JsonConfig.jsonConfig;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.comparesEqualTo;
+import static org.hamcrest.Matchers.is;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = KogitoSpringbootApplication.class)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class DMNScoreCardTest {
+@QuarkusTest
+public class DMNEquipmentMaintenanceTest {
 
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
+    static {
+        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
     }
 
     @Test
-    public void testEvaluateScoreCardDMN() {
-        String inputData = "{\"age\": 23.0, \"occupation\": \"SKYDIVER\", \"residenceState\": \"AP\", \"validLicense\": true}";
+    public void testEvaluateEquipmentMaintenanceDMN() {
+        String inputData = "{\"monthsInService\": 23.0, \"equipmentType\": \"PRESS\", \"operatingEnvironment\": \"OUTDOOR\", \"inspectionPassed\": true}";
         given()
                 .config(RestAssured.config().jsonConfig(jsonConfig().numberReturnType(JsonPathConfig.NumberReturnType.DOUBLE)))
                 .contentType(ContentType.JSON)
                 .body(inputData)
                 .when()
-                .post("/TestScoreCardDMN")
+                .post("/EquipmentMaintenance")
                 .then()
                 .statusCode(200)
-                .body("ScoreCardBKM", is("function ScoreCardBKM( age, occupation, residenceState, validLicense )"))
-                .body("age", is(comparesEqualTo(23))) // was input
-                .body("occupation", is(comparesEqualTo("SKYDIVER"))) // was input
-                .body("residenceState", is(comparesEqualTo("AP"))) // was input
-                .body("validLicense", is(comparesEqualTo(true))) // was input
-                .body("Score", is(comparesEqualTo(21.345)));// real decision output
+                .body("EquipmentMaintenanceScoreBKM", is("function EquipmentMaintenanceScoreBKM( monthsInService, equipmentType, operatingEnvironment, inspectionPassed )"))
+                .body("monthsInService", is(comparesEqualTo(23))) // was input
+                .body("equipmentType", is(comparesEqualTo("PRESS"))) // was input
+                .body("operatingEnvironment", is(comparesEqualTo("OUTDOOR"))) // was input
+                .body("inspectionPassed", is(comparesEqualTo(true))) // was input
+                .body("'Maintenance Priority'", is(comparesEqualTo(21.345))) // real decision output
+        ;
     }
 }
