+++ kogito-quarkus-examples/dmn-pmml-quarkus-example/src/test/java/org/kie/kogito/dmn/pmml/quarkus/example/ScoreCardTest.java
@@ -32,8 +32,8 @@ import static org.kie.kogito.dmn.pmml.quarkus.example.CommonTestUtils.testResult
 @QuarkusTest
 public class ScoreCardTest {
 
-    private static final String BASE_PATH = "/Testscorecard/SampleScore";
-    private static final String TARGET = "overallScore";
+    private static final String BASE_PATH = "/Testscorecard/MaintenancePriorityScore";
+    private static final String TARGET = "maintenancePriorityScore";
 
     static {
         RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
@@ -41,13 +41,13 @@ public class ScoreCardTest {
 
     @Test
     public void testEvaluateScoreCardResult() {
-        String inputData = "{\"age\": 23.0, \"occupation\": \"SKYDIVER\", \"residenceState\": \"AP\", \"validLicense\": true}";
+        String inputData = "{\"monthsInService\": 23.0, \"equipmentType\": \"PRESS\", \"operatingEnvironment\": \"OUTDOOR\", \"inspectionPassed\": true}";
         testResult(inputData, BASE_PATH, TARGET, 21.345f);
     }
 
     @Test
     public void testEvaluateScoreCardResultDescriptive() {
-        String inputData = "{\"age\": 23.0, \"occupation\": \"SKYDIVER\", \"residenceState\": \"AP\", \"validLicense\": true}";
+        String inputData = "{\"monthsInService\": 23.0, \"equipmentType\": \"PRESS\", \"operatingEnvironment\": \"OUTDOOR\", \"inspectionPassed\": true}";
         final Map<String, Object> expectedResultMap = Collections.singletonMap(TARGET, 21.345f);
         testDescriptive(inputData, BASE_PATH, TARGET, expectedResultMap);
     }
